This PR implements the use of quantized KV caches for the CPU backend when using FlashAttention. This is done via switch statements in `ggml_compute_forward_flash_attn_ext_f16`; from what I can tell this does not significantly affect performance. I also added comments and did some small performance optimizations such as converting Q only once to FP16/q8_0/q8_1 and using multiplications for scaling rather than divisions. The amount of `wdata` has been increased but only on the order of kiB which should be negligible. On my desktop with a Ryzen 5950X power limited to 95W the performance changes as follows:

| model         | backend | threads | fa | test          | t/s master | t/s PR | Speedup |
| ------------- | ------- | ------: | -: | ------------: | ---------: | -----: | ------: |
| llama 7B Q4_0 | CPU     |      16 |  1 | pp2048        |      33.43 |  35.29 |    1.06 |
| llama 7B Q4_0 | CPU     |      16 |  1 | tg512         |      10.59 |  10.71 |    1.01 |

When benchmarking, be mindful of the order in which you run the tests since a hot CPU will perform worse and this difference can easily be larger than the performance difference from this PR. I did a warmup run prior to the actual measurements.