@slaren 2024-05-18
It would be preferable to use the more generic function pointers in `type_traits` in the same way mul_mat does.