@JohannesGaessler 2024-05-19
I tested using the same FP32 code for both FP16 and quantized V cache but the performance was worse.